Ajustes de balance y feedback visual
Se ajusto el balance con la ballesta, y se añadió feedback visual al jugador cuando es golpeado

+++ Assets/Scripts/UI/HealthSlider.cs
@@ -18,19 +18,19 @@ public class HealthSlider : MonoBehaviour
         SetHealth(plantData.life);
     }
 
-    public void SetMaxHealth(int maxHealth)
+    public void SetMaxHealth(float maxHealth)
     {
         slider.maxValue = maxHealth;
     }
 
-    public void SetHealth(int health)
+    public void SetHealth(float health)
     {
         slider.value = health;
     }
 
     // Call this method in Boss script inserting health / LIFE argument.
     // In Start and update? Maybe calling them individualy insted of using this method. As you wish.
-    public void InitializeHealth(int health, int maxHealth)
+    public void InitializeHealth(float health, float maxHealth)
     {
         SetMaxHealth(maxHealth);
         SetHealth(health);
